pyproject.toml
[project]
name = "fiverr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
